+++ build.sbt
@@ -1,6 +1,6 @@
 name := "wikiClade"
 
-version := "1.1.0"
+version := "1.0.7"
 
 scalaVersion := "2.12.8"
