+++ src/main/scala/cladograms/Main.scala
@@ -24,20 +24,12 @@ object Main extends App {
   }
 
   def getLeafClade(name: String): Clade = {
-    WikiClade.newInputClade(name)
+    EnWikipediaClade(name, Some("/wiki/" + name.replaceAll(" ", "_")), 0)
   }
 
   def dotToGraphVizUrl(dot: String): String = {
     "https://dreampuf.github.io/GraphvizOnline/#" +
       URLEncoder.encode(dot, StandardCharsets.UTF_8.toString).replaceAll("\\+", " ")
-  }
-
-  def parseVerbosity(verbosityStr: String): Int = {
-    if (verbosityStr.startsWith("m")) 0
-    else if (verbosityStr.startsWith("l")) 20
-    else if (verbosityStr.startsWith("n")) 30
-    else if (verbosityStr.startsWith("c")) 100
-    else Try(verbosityStr.toInt).getOrElse(30)
   }
 
   def parseArgs: (List[String], Map[String, String]) = {
@@ -75,8 +67,8 @@ object Main extends App {
 
   val (inputs, configs) = parseArgs
   val verbosity = configs.get("-v") match {
-    case None => 30
-    case Some(s) => parseVerbosity(s)
+    case None => None
+    case Some(s) => Try(s.toInt).toOption
   }
 
   val format = configs.get("-f") match {
@@ -92,5 +84,5 @@ object Main extends App {
     case Some(s) => if (s.endsWith(format.fileExtension)) s else s + format.fileExtension
   }
 
-  renderer(inputs, Some(verbosity), format).toFile(new File(s"$outFile"))
+  renderer(inputs, verbosity, format).toFile(new File(s"$outFile"))
 }
