+++ src/main/scala/cladograms/WikiProxy.scala
@@ -8,14 +8,23 @@ import scala.util.Try
 
 object WikiProxy {
 
-  private val pageTitles: mutable.Map[String, Option[String]] = mutable.Map()
+  private val pageTitles: mutable.Map[String, String] = mutable.Map()
+  private val singleUseDocCache: mutable.Map[String, Document] = mutable.Map()
 
   private def getDoc(url: String): Option[Document] = {
+    if (singleUseDocCache contains url) {
+      singleUseDocCache remove url
+    } else {
+      //println(s"fetching $url")
       Try(Jsoup.connect(url).execute().parse()).toOption
+    }
   }
 
   private def storeTitle(url: String, doc: Document): Unit = {
-    pageTitles(url) = Try(doc.select("title").text()).toOption
+    Try(doc.select("title").text()).toOption match {
+      case Some(title) => pageTitles(url) = title
+      case None => ()
+    }
   }
 
 
@@ -24,9 +33,7 @@ object WikiProxy {
       case Some(doc) =>
         if (!(pageTitles contains url)) storeTitle(url, doc)
         Some(doc)
-      case None =>
-        pageTitles(url) = None
-        None
+      case None => None
     }
   }
 
@@ -34,12 +41,12 @@ object WikiProxy {
     if (!(pageTitles contains url)) {
       getDoc(url) match {
         case Some(doc) =>
+          singleUseDocCache(url) = doc
           storeTitle(url, doc)
-        case None =>
-          pageTitles(url) = None
+        case None => ()
       }
     }
-    pageTitles(url)
+    pageTitles.get(url)
   }
 
 }
